Refactor and bug fix of backtest
