+++ backtest/backtest_base.py
@@ -8,8 +8,8 @@ from typing import List, Dict
 import pandas as pd
 import numpy as np
 
-from backtest.event import Event, CashFlowChange, UpdatePortfolio
-from backtest.order import LimitOrder, FilledOrder, CanceledOrder
+from backtest.event import (Event, CashFlowChange, OptionAssigned, OptionExpired,
+                            UpdatePortfolio, LimitOrder, FilledOrder, CanceledOrder)
 from backtest.portfolio import Portfolio
 from .data_parser.ohlcv import OHLCV
 from .data_parser.option_chain import OptionChain
@@ -19,16 +19,17 @@ from .utils.instrument import InstrumentType
 
 
 class BacktestBase(object):
-    def __init__(self, history_data_path: Dict[str, str], frequency: FREQUENCY, start_date: str, end_date: str,
+    def __init__(self, history_data_path: str, instruments: Dict, frequency: FREQUENCY, start_date: str, end_date: str,
                  **kwargs) -> None:
         """
-        :param history_data_path: dictionary with symbol as keys and data paths as values.
+        :param history_data_path: path to directory that contains all market data. The path needs to be combined with symbol and instrument type.
         :param frequency: frequency of the backtest data.
         :param start_date: start date string in format YYYY-MM-DD and it's inclusive.
         :param end_date: end date string in format YYYY-MM-DD and it's inclusive.
         :param kwargs: other potential configs.
         """
         self.history_data_path = history_data_path
+        self.instruments = instruments
         self.start_date = start_date
         self.end_date = end_date
         initial_cash_balance = kwargs.get('initial_cash_balance', 0)
@@ -37,6 +38,7 @@ class BacktestBase(object):
         self.portfolio_snapshots: List[Dict] = []
         self.net_cash_flow = initial_cash_balance
         self.period_returns: List[float] = []
+        self._load_data()
 
     @staticmethod
     def get_simple_return(start_value: float, end_value: float) -> float:
@@ -87,19 +89,34 @@ class BacktestBase(object):
                 open_orders[event.order_id] = event
             elif isinstance(event, CanceledOrder):
                 del open_orders[event.order_id]
-            # TODO: add events of option assign and option expire.
+            elif isinstance(event, OptionAssigned):
+                self.portfolio.option_assigned(event)
+            elif isinstance(event, OptionExpired):
+                self.portfolio.option_expired(event)
             elif isinstance(event, CashFlowChange):
                 if last_value:
                     prices = {}
                     for symbol in self.portfolio.positions.keys():
-                        end_of_day = event.ts.normalize() + pd.Timedelta(days=1) - pd.Timedelta(microseconds=1)
-                        intraday_prices = self.ohlcv_data[symbol].get_ohlcv_by_date_string(event.ts, end_of_day)
-                        try:
-                            close_price = intraday_prices.iloc[-1].close
-                        except IndexError:
-                            print(f"{event.ts} data not exist")
-                            continue
-                        prices[symbol] = close_price
+                        position = self.portfolio.positions[symbol]
+                        instrument = position.instrument
+                        close_price = None
+
+                        if instrument.type == InstrumentType.STOCK:
+                            # FIXME later: Here it assumes price frequency is hourly. However, this method should be encapsulated to the DataParse to
+                            # accommodate for all frequency. Currently we only have hourly stock data.
+                            end_of_day = event.ts.normalize() + pd.Timedelta(days=1) - pd.Timedelta(microseconds=1)
+                            intraday_prices = self.ohlcv_data[symbol].get_ohlcv_by_date_string(event.ts, end_of_day)
+                            try:
+                                close_price = intraday_prices.iloc[-1].close
+                            except IndexError:
+                                print(f"{event.ts} {symbol} data not exist")
+                                continue
+                        elif instrument.type == InstrumentType.OPTION:
+                            underlying_symbol = instrument.underlying_symbol
+                            option_chain_data = self.option_data[underlying_symbol]
+                            close_price = option_chain_data.get_instrument_price(event.ts.date(), instrument)
+                        if close_price is not None:
+                            prices[symbol] = close_price
                     self.portfolio.update_portfolio(prices)
                     period_return = self.get_simple_return(last_value, self.portfolio.portfolio_value)
                     self.period_returns.append(period_return)
@@ -115,12 +132,11 @@ class BacktestBase(object):
                 raise ValueError(f"Invalid event type: {type(event)}.")
 
     def _load_data(self) -> None:
-        instruments = self.config.get("instrument", {})
-        if InstrumentType.OPTION.value in instruments:
-            self.ohlcv_data = self._load_option_data(instruments[InstrumentType.OPTION.value])
+        if InstrumentType.OPTION.value in self.instruments:
+            self.option_data = self._load_option_data(self.instruments[InstrumentType.OPTION.value])
 
-        if InstrumentType.STOCK.value in instruments:
-            self.option_data = self._load_stock_data(instruments[InstrumentType.STOCK.value])
+        if InstrumentType.STOCK.value in self.instruments:
+            self.ohlcv_data = self._load_stock_data(self.instruments[InstrumentType.STOCK.value])
 
     def _load_stock_data(self, symbols: List[str]) -> Dict[str, OHLCV]:
         """
